+++ index.html
@@ -50,6 +50,6 @@
    
   </main>
   <footer>Copyright...</footer>
-  <script type="module" src="index.js"></script>
+  <script src="app.js"></script>
   </body>
 </html>
